rozdělení kartiček a generování seznamu
rozdělení kartiček a generování seznamu
