+++ client/src/pages/RecipePage.js
@@ -3,7 +3,7 @@ import { Button } from "react-bootstrap";
 import { useSearchParams } from "react-router-dom";
 import AddRecipe from "../components/AddRecipe";
 import AdvancedSearch from "../components/AdvancedSearch";
-import MenuCardsOutput from "../components/MenuCardsFilter";
+import MenuCardsOutput from "../components/MenuCardsGrid";
 
 const RecipePage = () => {
     const [searchParams] = useSearchParams();
@@ -24,7 +24,7 @@ const RecipePage = () => {
             <h1 className='container-fluid d-flex justify-content-center'>Recipe Page</h1>
 
             <Button onClick={() => {setShowSearch(!showSearch)}} className='pt-1 float-end'>
-                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-filter" viewBox="0 0 16 16">
+                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" className="bi bi-filter" viewBox="0 0 16 16">
                     <path d="M6 10.5a.5.5 0 0 1 .5-.5h3a.5.5 0 0 1 0 1h-3a.5.5 0 0 1-.5-.5zm-2-3a.5.5 0 0 1 .5-.5h7a.5.5 0 0 1 0 1h-7a.5.5 0 0 1-.5-.5zm-2-3a.5.5 0 0 1 .5-.5h11a.5.5 0 0 1 0 1h-11a.5.5 0 0 1-.5-.5z"/>
                 </svg>
             </Button>
